+++ CHANGES.rst
@@ -4,6 +4,8 @@
 Bug Fixes
 ---------
 
+- Change box zoom to always maintain aspect ratio. [#1726]
+
 Cubeviz
 ^^^^^^^